Split context opening/closing and variable replacing.
With this, we are now able to replace compound literals used as
function arguments, inside the condition of if statements, and a
number of other ways of using compound literals. The main use of
compound literals that we can't correctly convert yet, at this point,
is its use in conditional statements (a ? b : c).