Fix conditional statements.
We can now convert all uses of compound literals in the unit test. There
is some corner cases that probably don't work correctly, e.g. where order
of statements or the fact that it's executed or not matters (since we do
all that unconditionally), but at least most of it is there.