Add a compiler frontend for the converter
This parses the compiler flags, and if the command is supposed to
compile, then preprocess, convert and compile in three steps, with
command line flags mapped to the two process calls.

Use with configure --cc="converter gcc" (for initial testing)
or --cc="converter cl" (for the actual intended use, might need some
more parameter handling tweaking).