Handle initializing unions in functions with a separate assignment call
This avoids having to convert floats to their binary representation,
for runtime assignments.

This converts code like

void func() {
    union foo a = { .f = f };
    int a;
    ...
}

into

void func() {
    union foo a;
    a.f = f;
    {
        int a;
        ....
    }
}

In practice it isn't formatted this prettily though. The closing
extra braces are placed on separate rows, which probably will mess
up the row numbering (for debuggers etc). It should ideally be grouped
close to the existing brace that close the function/scope, but with
my current implementation, it is prepended to the following token
instead. (And in that case, we need the extra newline, since
if the following line starts with a #, it will be messed up if the
extra } is prepended there - that's why the current extra newline
is there.)